requirements.txt
asgiref==3.5.0
boto3==1.24.49
botocore==1.27.49
dj-database-url==0.5.0
Django==4.0.3
django-cors-headers==3.12.0
django-filter==22.1
django-ses==3.1.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
gunicorn==20.1.0
jmespath==1.0.1
psycopg2-binary==2.9.3
PyJWT==2.4.0
python-dateutil==2.8.2
pytz==2022.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.11
whitenoise==6.2.0
